+++ modules/blogging/src/Volo.Blogging.Web/Pages/Blogs/BloggingPageModel.cs
@@ -1,10 +1,11 @@
 ﻿using System;
+using System.Linq;
 using System.Threading.Tasks;
 using Volo.Abp.AspNetCore.Mvc.UI.RazorPages;
 using Volo.Blogging.Blogs;
 using Volo.Blogging.Blogs.Dtos;
 
-namespace Volo.Blogging.Pages.Blog
+namespace Volo.Blogging.Pages.Blogs
 {
     public abstract class BloggingPageModel : AbpPageModel
     {
@@ -15,7 +16,7 @@ namespace Volo.Blogging.Pages.Blog
         
         protected async Task<BlogDto> GetBlogAsync(IBlogAppService blogAppService, BloggingUrlOptions blogOptions, string blogShortName)
         {
-            if (!blogOptions.SingleBlogMode.Enable)
+            if (!blogOptions.SingleBlogMode.Enabled)
             {
                 return await blogAppService.GetByShortNameAsync(blogShortName);
             }
@@ -26,12 +27,7 @@ namespace Volo.Blogging.Pages.Blog
             }
 
             var blogs = await blogAppService.GetListAsync();
-            if (blogs.Items.Count == 1)
-            {
-                return blogs.Items[0];
-            }
-            
-            return await blogAppService.GetByShortNameAsync(blogOptions.SingleBlogMode.BlogName);
+            return blogs.Items.Count == 1 ? blogs.Items[0] : blogs.Items.SingleOrDefault(x => x.ShortName == blogShortName);
         }
     }
 }
