+++ modules/blogging/src/Volo.Blogging.Web/Pages/Blogs/Index.cshtml
@@ -1,9 +1,8 @@
 ﻿@page
-@using Volo.Blogging.Pages.Blog
 @using Microsoft.AspNetCore.Mvc.Localization
 @using Volo.Blogging.Localization
 @inject IHtmlLocalizer<BloggingResource> L
-@model IndexModel
+@model Volo.Blogging.Pages.Blogs.IndexModel
 @{
     ViewBag.PageTitle = "Blogs";
 }
