+++ modules/blogging/src/Volo.Blogging.Web/Pages/Blogs/Posts/New.cshtml
@@ -1,11 +1,11 @@
 ﻿@page
 @using Volo.Abp.AspNetCore.Mvc.UI.Packages.TuiEditor
-@using Volo.Blogging.Pages.Blog.Posts
 @using Volo.Blogging.Posts
 @using Microsoft.AspNetCore.Mvc.Localization
 @using Volo.Blogging.Localization
+@using Volo.Blogging.Pages.Blogs.Posts
 @inject IHtmlLocalizer<BloggingResource> L
-@model NewModel
+@model Volo.Blogging.Pages.Blogs.Posts.NewModel
 @{
     ViewBag.PageTitle = @L["NewBlogPost"];
 }
